first phase of transition to unified Unicode build: 1. changed c_str() to return wxCStrData (implicitly convertible to wxChar*) 2. added template type-safe wrappers for vararg functions 3. added wxUniChar class representing single Unicode character 4. changed wxString::operator[] and wxString::iterator to return wxUniChar
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@44865 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/unix/joystick.cpp
@@ -394,7 +394,7 @@ int wxJoystick::GetNumberButtons() const
     if (m_device != -1)
         ioctl(m_device, JSIOCGBUTTONS, &nb);
 
-    if (nb > wxJS_MAX_BUTTONS)
+    if ((int)nb > wxJS_MAX_BUTTONS)
         nb = wxJS_MAX_BUTTONS;
 
     return nb;
@@ -407,7 +407,7 @@ int wxJoystick::GetNumberAxes() const
     if (m_device != -1)
         ioctl(m_device, JSIOCGAXES, &nb);
 
-    if (nb > wxJS_MAX_AXES)
+    if ((int)nb > wxJS_MAX_AXES)
         nb = wxJS_MAX_AXES;
 
     return nb;
